+++ packages/wonder-blocks-button/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @khanacademy/wonder-blocks-button
 
+## 3.0.10
+
+### Patch Changes
+
+-   Updated dependencies [b561425a]
+-   Updated dependencies [a566e232]
+-   Updated dependencies [d2b21a6e]
+    -   @khanacademy/wonder-blocks-core@4.6.0
+    -   @khanacademy/wonder-blocks-clickable@2.4.2
+    -   @khanacademy/wonder-blocks-icon@1.2.33
+    -   @khanacademy/wonder-blocks-progress-spinner@1.1.37
+    -   @khanacademy/wonder-blocks-typography@1.1.35
+
 ## 3.0.9
 
 ### Patch Changes