+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 2.1.37
+
+### Patch Changes
+
+-   Updated dependencies [b561425a]
+-   Updated dependencies [a566e232]
+-   Updated dependencies [d2b21a6e]
+    -   @khanacademy/wonder-blocks-core@4.6.0
+    -   @khanacademy/wonder-blocks-typography@1.1.35
+
 ## 2.1.36
 
 ### Patch Changes